refactor(server): move route support modules into package
